COM-3913 - define generateCompletionCertificate in composable
