+++ src/core/composables/completionCertificates.js
@@ -1,9 +1,11 @@
 import { ref } from 'vue';
 import CompletionCertificates from '@api/CompletionCertificates';
-import { NotifyNegative } from '@components/popup/notify';
+import { NotifyNegative, NotifyPositive } from '@components/popup/notify';
+import { GENERATION } from '@data/constants';
 
-export const useCompletionCertificates = (completionCertificates, monthOptions = []) => {
+export const useCompletionCertificates = (completionCertificates, refreshCompletionCertificates, monthOptions = []) => {
   const tableLoading = ref(false);
+  const disableButton = ref(false);
 
   const getCompletionCertificates = async (query) => {
     try {
@@ -19,10 +21,27 @@ export const useCompletionCertificates = (completionCertificates, monthOptions =
     }
   };
 
+  const generateCompletionCertificate = async (completionCertificateId) => {
+    try {
+      disableButton.value = true;
+      await CompletionCertificates.update(completionCertificateId, { action: GENERATION });
+      NotifyPositive('Certificat de réalisation généré.');
+
+      await refreshCompletionCertificates();
+    } catch (e) {
+      console.error(e);
+      NotifyNegative('Erreur lors de la génération du certificat.');
+    } finally {
+      disableButton.value = false;
+    }
+  };
+
   return {
     // Data
     tableLoading,
+    disableButton,
     // Methods
     getCompletionCertificates,
+    generateCompletionCertificate,
   };
 };
